stegothreat/requirements.txt
streamlit==1.32.0
pillow==10.1.0
numpy==1.24.3
yara-python==4.5.1
requests==2.31.0